requirements.txt
minio==7.0.0
urllib3==1.26.7
python-dotenv==0.19.0